move util function in utils files
